requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
Django==2.2.2
django-cors-headers==3.0.2
django-oauth-toolkit==1.2.0
djangorestframework==3.9.4
geoip2==2.9.0
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
maxminddb==1.4.1
numpy==1.16.4
oauthlib==3.0.2
pandas==0.24.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2-binary==2.8.2
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.3
wcwidth==0.1.7
